playing with torch and tensorboard, specifically:
*contrastive learning* with positive and negative examples from CA images,
which should eventually grow an encoder for embeddings of CA patterns
that can be somehow useful later on, e.g. for classification of rules.

However, much wasted *compute* on the way..

Also tried pytorch-lightning which is basically acceptable but my implementation
of an on-the-fly CA dataset using the generic torch IterableDataset and DataLoader
framework had terrible runtime for generating the CA images, so i dropped the experimentation
with lighting for the moment and went back to a self-hacked training framework,
while keeping tensorboard on board.